fix: inline vitest module augmentation into vitest.mts
Move the `declare module 'vitest'` augmentation from the standalone
src/vitest.types.ts into src/vitest.mts so that the `./vitest` subpath
export no longer ships an empty, unused CommonJS file (vitest.types.js).
The published types entry now points to the .d.mts emitted alongside
the runtime, keeping types and runtime co-located.
