+++ CODEMAP.md
@@ -5,10 +5,10 @@ Generated by `npm run codemap`. This file is intentionally machine-readable for 
 ````json
 {
     "schema": "codemap.v1",
-    "generatedAt": "2026-04-30T16:43:46.454Z",
+    "generatedAt": "2026-05-12T14:43:34.440Z",
     "package": {
         "name": "png-visual-compare",
-        "version": "6.0.0",
+        "version": "6.1.0",
         "description": "Node.js utility to compare PNG files or their areas",
         "keywords": ["visual testing", "compare png", "compare images", "png", "png compare", "png diff", "image diff", "diff"],
         "homepage": "https://github.com/dichovsky/png-visual-compare#readme",
@@ -32,7 +32,7 @@ Generated by `npm run codemap`. This file is intentionally machine-readable for 
                 "default": "./out/index.js"
             },
             "./vitest": {
-                "types": "./out/vitest.types.d.ts",
+                "types": "./out/vitest.d.mts",
                 "import": "./out/vitest.mjs",
                 "default": "./out/vitest.mjs"
             },
@@ -57,7 +57,7 @@ Generated by `npm run codemap`. This file is intentionally machine-readable for 
             "lint": "eslint .",
             "lint:fix": "eslint . --fix",
             "prepublishOnly": "npm test && npm run build",
-            "test": "npm run test:unit & npm run test:e2e",
+            "test": "npm run test:unit && npm run test:e2e",
             "test:docker": "npm run clean && npm run docker:build && npm run docker:run",
             "pretest:e2e": "npx playwright install chromium",
             "test:e2e": "playwright test",
@@ -72,16 +72,16 @@ Generated by `npm run codemap`. This file is intentionally machine-readable for 
             "pngjs": "~7.0.0"
         },
         "devDependencies": {
-            "@playwright/test": "~1.59.1",
-            "@types/node": "~25.6.0",
+            "@playwright/test": "~1.60.0",
+            "@types/node": "~25.7.0",
             "@types/pngjs": "~6.0.5",
-            "@vitest/coverage-v8": "~4.1.5",
-            "eslint": "~10.2.1",
+            "@vitest/coverage-v8": "~4.1.6",
+            "eslint": "~10.3.0",
             "prettier": "~3.8.3",
             "rimraf": "~6.1.3",
             "typescript": "~6.0.3",
-            "typescript-eslint": "~8.59.1",
-            "vitest": "~4.1.5"
+            "typescript-eslint": "~8.59.3",
+            "vitest": "~4.1.6"
         },
         "peerDependencies": {
             "jest": ">=29 <31",
@@ -1099,10 +1099,6 @@ Generated by `npm run codemap`. This file is intentionally machine-readable for 
                     "declaration": "export function validatePixelmatchOptions(opts: PixelmatchOptions): void { if (typeof opts !== 'object' || opts === null || Array.isArray(opts)) { throw new InvalidInputError('pixelmatchOptions must be an object when provided'); } validateUnitInterval('pixelmatchOptions.threshold', opts.threshold); validateUnitInterval('pixelmatchOptions.alpha', opts.alpha); validateBoolean('pixelmatchOptions.includeAA', opts.includeAA); validateBoolean('pixelmatchOptions.diffMask', opts.diffMask); if (opts.aaColor !== undefined) validateColorTuple('pixelmatchOptions.aaColor', opts.aaColor); if (opts.diffColor !== undefined) validateColorTuple('pixelmatchOptions.diffColor', opts.diffColor); if (opts.diffColorAlt !== undefined) validateColorTuple('pixelmatchOptions.diffColorAlt', opts.diffColorAlt); }"
                 }
             ]
-        },
-        {
-            "path": "src/vitest.types.ts",
-            "symbols": []
         }
     ]
 }
